use size_t types for moobinary<->raw_bytes, recode_chars

+++ net_multi.c
@@ -223,7 +223,7 @@ static int
 push_output(nhandle * h)
 {
     text_block *b;
-    int count;
+    ssize_t count;
 
     if (h->output_lines_flushed > 0) {
 	char buf[100];
@@ -265,7 +265,7 @@ static int
 pull_input(nhandle * h)
 {
     Stream *s = h->input;
-    int count;
+    ssize_t count;
     char buffer[1024];
     char *ptr, *end;
